Workaround for libyuv libjingle breakage.
libjingle depends on ConvertFromI420. This was previously available
through vplib. libjingle still has access to the vplib header, but the
implementation is no longer built.

Fortunately, the libyuv wrapper can supply the implementation, if we
hack the signature to return to the unsigned int types. We'll remove
this once libjingle has been updated to use libyuv directly.

Also, roll libyuv to r100 which fixes a gyp warning on Windows.

TEST=build

Review URL: http://webrtc-codereview.appspot.com/323004

git-svn-id: http://webrtc.googlecode.com/svn/trunk@1151 4adac7df-926f-26a2-2b94-8c16560cd09d