Fixed the sanity_check and started using the new webrtc_test.html file. Added capability for xvfb testing.
The purpose for the xvfb mode is to be able to run tests on the windowless environment on the Chromebot. Given the right input video, we can then write a relatively simple algorithm to analyze the screenshots and thereby conclude that video is playing.

BUG=
TEST=

Review URL: https://webrtc-codereview.appspot.com/447004

git-svn-id: http://webrtc.googlecode.com/svn/trunk@1890 4adac7df-926f-26a2-2b94-8c16560cd09d