Fixed a bug in the C code in a filterbank function in iSAC-fix, and further optimized ARM code.
The bug was introduced in the last CL during optimization.

With the new inine ARM assembly, the resulted assembly code of the loop is 28 instructions, compared to 47 lines by the original.

The optimized (with WEBRTC_ARCH_ARM_V7A defined in This file) and un-optimized code (with WEBRTC_ARCH_ARM_V7A not defined in This file) are bit-exact.
Review URL: https://webrtc-codereview.appspot.com/748004

git-svn-id: http://webrtc.googlecode.com/svn/trunk@2651 4adac7df-926f-26a2-2b94-8c16560cd09d