Fix audio_e2e_test command line arguments
The changes in r1068 that moved over to the webrtc/test/buildbot_tests.py launch script was not properly tested on the real machine for the audio_e2e_test. Due to that it contained a few syntax errors and paths that were not resolved as expected. This CL fixes this and has been tested more thorougly.

BUG=none
TEST=Ran, standing in the checkout dir:
out/Release/buildbot_tests.py -t audio_e2e_test
with successful result.

Review URL: https://webrtc-codereview.appspot.com/1070012

git-svn-id: http://webrtc.googlecode.com/svn/trunk@3438 4adac7df-926f-26a2-2b94-8c16560cd09d