Final round of LSan suppressions.
Since we don't have a trybot for LSan, this is the only
way to resolve errors only happening on the bots
(except running at the bot themselves).

I suspect some errors only occurs on the bots since
the compiler version differs (4.6 on bots, 4.6.3 on
my machine).

This last suppressions round was tested on an actual
Chrome infra VM as well, with no errors.

BUG=2528
TEST=ran successfully on the bot.
TBR=phoglund@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/2424004

git-svn-id: http://webrtc.googlecode.com/svn/trunk@4991 4adac7df-926f-26a2-2b94-8c16560cd09d