Revert "Add h.264 AVC SPS parsing for resolution."
The Mac64 Debug builder is broken for an unknown failure (trybot is
green, no failure obvious in the commit break). Reverting this CL to see
if it goes green again, and then relanding to see if it is just some
weird flaky build issue.

This reverts commit 077fa125a82ad2cddd63e43d4be8cbd54ee8ac5b.

BUG=
TBR=rollback

Review URL: https://webrtc-codereview.appspot.com/47019004

Cr-Commit-Position: refs/heads/master@{#9074}