Removing AudioCoding class, a.k.a the new ACM API
We have decided not to do a switch from old (AudioCodingModule) to new
(AudioCoding) API. Instead, we will gradually evolve the old API to
meet the new design goals.

As a consequence of this decision, the AudioCoding and AudioCodingImpl
classes are deleted. Also removing associated unit test sources. No
test coverage is lost with this operation, since the tests for the
"old" API are testing more than the deleted tests did.

BUG=webrtc:3520

Review URL: https://codereview.webrtc.org/1415163002

Cr-Commit-Position: refs/heads/master@{#10406}