GN: Fix dependency for rtc_base_approved
In https://codereview.webrtc.org/1921463002/ a dependency
on Foundation.framework was added to rtc_base_approved for the
GYP build but not for the GN build. This was fixed for some
targets in https://codereview.webrtc.org/1919673002/ but the proper
fix is to introduce an all_dependent_configs entry for rtc_base_approved.

NOTRY=True

Review URL: https://codereview.webrtc.org/1921513002

Cr-Commit-Position: refs/heads/master@{#12483}