GN: Refactor webrtc/base/BUILD.gn
Move the all_dependent_configs configurations into
a single config for better readability.

Sync the Mac and iOS frameworks between the GYP and the GN
build. Many entries that used to be default included in
https://code.google.com/p/chromium/codesearch#chromium/src/build/config/BUILD.gn
are no longer implicitly added for Mac, so to ensure they are
they're added back for the Mac build.

NOTRY=True

Review URL: https://codereview.webrtc.org/1914913002

Cr-Commit-Position: refs/heads/master@{#12484}