Remove the rtc_relative_path GYP variable and similar defines
This is a reland of https://codereview.webrtc.org/1903553003/ but with
the SRTP changes removed, since they're needed downstream.

The defines that can be used to alter the include paths for Expat and gtest
are no longer needed in WebRTC or Chromium. Remove them to simplify GYP.

Removed defines:
EXPAT_RELATIVE_PATH
GTEST_RELATIVE_PATH

They're all set in the Chromium build so this shouldn't affect Chromium:
https://code.google.com/p/chromium/codesearch#chromium/src/third_party/libjingle/libjingle.gyp

BUG=webrtc:4256
NOTRY=True
NOPRESUBMIT=True
TBR=perkj@webrtc.org

Review-Url: https://codereview.webrtc.org/1925733002
Cr-Commit-Position: refs/heads/master@{#12536}