Limit the maximum and minimum gain the IntelligibilityEnhancer can apply
Although the algorithm should generate the optimal gains, these limits are in place so that it doesn't go bananas in extreme noise situations or if there are some errors on the noise estimate. The limits were chosen as the extreme values from recordings with -3dB SNR, which is the minimum where the POLQA metric increases when processed.

Review-Url: https://codereview.webrtc.org/1925933002
Cr-Commit-Position: refs/heads/master@{#12550}